# Misc fixes, including timing