--- a/onewire.c
+++ b/onewire.c
@@ -69,7 +69,7 @@ volatile uint8_t cbuf;  // char buffer, current byte to be (dis)assembled
 #define OWT_RESET_PRESENCE (T_(40)-1)
 #define OWT_PRESENCE (T_(160)-1)
 #define OWT_READLINE (T_(30)-2)
-#define OWT_LOWTIME (T_(50)-2)
+#define OWT_LOWTIME (T_(40)-2)
 
 #if (OWT_MIN_RESET>240)
 #error Reset timing is broken, your clock is too fast