--- a/world.cfg
+++ b/world.cfg
@@ -229,18 +229,18 @@ defaults:
         8: 8
         9: 9
       adc:
-       adc:
-        1: 0
-        2: 1
-        3: 2
-        4: 3
-        5: 4
-        6: 5
-        7: 6
-        8: 7
-        9: T-
-        10: R
-        11: G
+        adc:
+          1: 0
+          2: 1
+          3: 2
+          4: 3
+          5: 4
+          6: 5
+          7: 6
+          8: 7
+          9: T-
+          10: R
+          11: G
     m88f:
       _doc: version of m88 with external crystal
       _ref: defaults.target.m88