Hi,

I've now successfully built the DS2423 device and got it working on a Mega88A with 8MHz crystal.
I've made some minor fixes to the build system, which I *think* are correct?

There was also some timing issues, the OWT_LOWTIME was too high, search failed with DS2480B (but worked with LinkUSB; see commit for more details).
Which master have you used?

Johan