@stromnet 2015-05-26
For the record; some timings:

```
Master: DS2480B
Slave: OWT_LOWTIME T_(50)-2

First bit in search; master read:
	1.47 uS  low-edge and int0-dbg (enter)
	10.9uS low-edge and int0-dbg (exit)
	(line high)
	10.3uS  low-edge and high-edge (device write-1)
	58.7uS low-edge and OWT_LOWTIME timer dbg-enter
	65.3uS low-edge and OWT_LOWTIME timer dbg-exit
	67.1 uS timer exit and start of next slot (low-edge)

	66.7uS  low-edge and next low-edge

Second bit in search; master-read:
	4.17 uS low-edge and int0-dbg-enter
	13.7 uS low-edge and int0-dbg-exit
	(line pulled low)
	61.7us low-edge and OWT_LOWTIME timer dbg-enter
	66.1uS low-edge and line release (Barely visible; timeslot is over!)
	68.1uS low-edge and OWT_LOWTIME timer dbg-exit

	67.3 uS  low-edge and next low-edge

<SLOT MISSED, THIRD BIT READ IS DONE 1 SLOT TOO LATE>

Third bit in search; master-write-decision:
	2.9 uS low-edge and int0-dbg-enter
	11.7uS low-edge and int0-dbg-exit
	39.5us low-edge and OWT_READLINE timer dbg-enter (sample 1)
	48.1us low-edge and OWT_READLINE timer dbg-exit

	67.3 us low-edge and next low-edge


Tested without debug pin, 1w looked same.
```
```
Master: DS2480B
Slave: OWT_LOWTIME T_(40)-2

First bit in search; master read:
	2.9uS  low-edge and int0-dbg (enter)
	11.9uS low-edge and int0-dbg (exit)
	(line high)
	11.9uS  low-edge and high-edge (device write-1)
	51.5uS low-edge and OWT_LOWTIME timer dbg-enter
	58.3uS low-edge and OWT_LOWTIME timer dbg-exit
	8.1 uS timer exit and start of next slot (low-edge)

	67uS  low-edge and next low-edge

Second bit in search; master-read:
	2.1uS low-edge and int0-dbg-enter
	11.5uS low-edge and int0-dbg-exit
	(line pulled low)
	52.1us low-edge and OWT_LOWTIME timer dbg-enter
	56.7uS low-edge and line release
	58.1uS low-edge and OWT_LOWTIME timer dbg-exit

	67uS  low-edge and next low-edge


Third bit in search; master-write-decision:
	2.9uS low-edge and int0-dbg-enter
	11.9uS low-edge and int0-dbg-exit
	40.5us low-edge and OWT_READLINE timer dbg-enter (sample 1)
	47.9us low-edge and OWT_READLINE timer dbg-exit

	67.1 us low-edge and next low-edge

```
```
Master:LinkUSB
Slave: OWT_LOWTIME T_(50)-2

First bit in search; master read:
	3.6 uS  low-edge and int0-dbg (enter)
	12.2 uS low-edge and int0-dbg (exit)
	(line high)
	9.0uS  low-edge and high-edge (device write-1)
	60.2uS low-edge and OWT_LOWTIME timer dbg-enter
	67.4uS low-edge and OWT_LOWTIME timer dbg-exit
	
	80.4uS  low-edge and next low-edge

Second bit in search; master-read:
	2.02 uS low-edge and int0-dbg-enter
	11.0 uS low-edge and int0-dbg-exit
	(line pulled low)
	59.2us low-edge and OWT_LOWTIME timer dbg-enter
	63.4uS low-edge and line release
	66.8uS low-edge and OWT_LOWTIME timer dbg-exit

	89.2 uS  low-edge and next low-edge

Third bit in search; master-write-decision:
	3.42uS low-edge and int0-dbg-enter
	12.2uS low-edge and int0-dbg-exit
	41us low-edge and OWT_READLINE timer dbg-enter (sample 1)
	48.2us low-edge and OWT_READLINE timer dbg-exit

	83 us low-edge and next low-edge

```