@stromnet 2015-05-27
Doh, my bad. I've changed my upstreams to m-o-a-t but for some reason I picked this repo when doing the pull request.

Thanks for merge!
Have you made any measurements on 8MHz internal RC? From the data-sheet, it should run at approx 8Mhz at 3V. At 5.0V it is slightly faster, I think, which could make the old values work better.