@smurfix 2015-05-27
On 27.05.2015 07:58, Johan Ström wrote:
> Have you made any measurements on 8MHz internal RC? From the
> data-sheet, it should run at approx 8Mhz at 3V. At 5.0V it is slightly
> faster, I think, which could make the old values work better. 
Not yet; I'll test on my hardware later this week, too much else to do
at the moment :-/

I'm doing 8MHz at 5V due to sheer laziness on my part (don't want to
bother soldering the external crystal if I can avoid it).

-- 
-- Matthias Urlichs