@stromnet 2015-05-27
Did some additional testing, with internal 8MHz @ 5V, resulting in some thoughts on changing the timings:

```
Reset pulse (clean restarted slave):

1.2uS falling-edge to int-enter  (OWM_SLEEP)
10.2uS edge to int-exit
422uS edge to timer-min-reset enter
442uS edge to timer-min-reset exit
510uS falling-edge to rising-edge
53uS rising-edge to OWT_RESET_PRESENCE-timer
56.4uS rising-edge to our falling edge (presence pulse)
166uS, falling edge to rising edge (lenght of our presence pulse; OWT_PRESENCE)
```
Seems legit!

In a non-first-transaction test I also see a DBG pulse 40uS after first low edge, which I guess is a OWT_READLINE, followed by a reset-sense-pulse at 436us after low edge. Then it continues as above after the rising edge.

Looking at the search-bit-pattern:
```
read-slot-1 (slave writes 1)
  2.6us after low edge, int enter
  11us after low edge, int exit
  52.4uS after low edge, write1-timer-enter
  58.8us after low edge, write1-timer-exit

read-slot-2 (slave writes 0)
  2.8us after low edge, int enter
  11us after low edge, int exit
  (line brought low)
  51.2uS after low edge, write0-timer-enter
  55.2 us after low edge, line released
  58.6us after low edge, write0-timer-exit

master-write-slot-3 (slave reads)
  2us after low edge, int enter
  12.2us after low edge, int exit
  40uS after low edge, sample-timer-enter
  47.8us after low edge, sample-timer-exit
```

I'd say thats pretty much the same as with the 8MHz crystal, and within limits.

But, OWT_LOWTIME might be a bit high, leaving the ISR 58uS after initial low edge. Given a slot width of 60uS, with >5uS recovery time (1), this gives us ~7uS margin until the next slot. The master should sample somewhere around 11-15uS after low-edge (2), so 58uS is quite high. 
In comparison, a DS18B20 keeps the line low for 27.4uS during the search.

With `LOWTIME (T_(40)-2)`, the line is kept low for 58uS (above)
With `LOWTIME (T_(30)-2)`, the line is kept low for 43uS.
With `LOWTIME (T_(25)-2)`, the line is kept low for 39uS.
With `LOWTIME (T_(20)-2)`, the line is kept low for 35uS.

The device is found fine in either of the above, in my test bench. I'm not sure how much a more complex network would impact, but 35uS seems quite sufficient, compared to the others? The main upside with a lower timeout is for parasite powered devices in the net, with maximized pull-up time.

What are your previous experience from tweaking the timings?

--

1) according to DS2406 spec; cannot find a PDF with the "definitive" answer? The DS2480B does not have any spec on recovery time.. 

2) The DS2480B spec specifies Standard sampling at 8+3uS = 11uS. The DS2406 pdf says "The optimal sampling point for the master is as close as possible to the end time of the 15μs tRDV period without exceeding tRDV. For the case of a Read-one time slot, this maximizes the amount of time for the pull-up resistor to recover the line to a high level. For a Read-zero time slot it ensures that a read will occur before the fastest 1-Wire devices(s) release the line (tRELEASE = 0)."